#115 test: import name refactoring
