--- src/pages/CommunityPage.jsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-import CommunityLanding from "@/components/Community/CommunityLanding/CommunityLanding";
-
-const CommunityPage = () => <CommunityLanding />;
-
-export default CommunityPage;
